nvme: move fw commands into a plugin
There are too many top-level commands cluttering the space. Move the
firmware commands into a plugin, e.g.,

  nvme fw-commit -> nvme fw commit
