+++ meson_options.txt
@@ -231,7 +231,7 @@ option(
   'plugins',
   type : 'array',
   choices: ['amzn', 'config', 'dapustor', 'dell', 'dera', 'dir',
-            'exclusion', 'fdp', 'feat', 'huawei', 'ibm', 'id', 'innogrit', 'inspur',
+            'exclusion', 'fdp', 'feat', 'fw', 'huawei', 'ibm', 'id', 'innogrit', 'inspur',
             'intel', 'io-mgmt', 'keys', 'lm', 'log', 'mangoboost', 'memblaze', 'micron', 'nbft',
             'netapp', 'ns', 'nvidia', 'nvme-mi', 'ocp', 'registry', 'resv', 'sandisk', 'scaleflux',
             'seagate', 'security', 'sed', 'shannon', 'solidigm', 'ssstc', 'toshiba',
